Rename import statements for Girder 3
