+++ docker/requirements.txt
@@ -1,88 +1,100 @@
-amqp==2.4.2
+amqp==2.5.0
+apipkg==1.5
 asn1crypto==0.24.0
+atomicwrites==1.3.0
 attrs==19.1.0
 backports.functools-lru-cache==1.5
-bcrypt==3.1.6
+bcrypt==3.1.7
 billiard==3.6.0.0
-boto3==1.9.130
-botocore==1.12.130
-cachetools==3.1.0
-celery==4.3.0
-certifi==2019.3.9
-cffi==1.12.2
+boto3==1.9.174
+botocore==1.12.174
+cachetools==3.1.1
+celery==4.4.0rc2
+certifi==2019.6.16
+cffi==1.12.3
 chardet==3.0.4
-cheroot==6.5.4
+cheroot==6.5.5
 CherryPy==11.0.0
 Click==7.0
 click-plugins==1.1.1
-cryptography==2.6.1
+coverage==5.0a5
+cryptography==2.7
 cycler==0.10.0
 decorator==4.4.0
 diskcache==3.1.1
-docker==3.7.2
-docker-pycreds==0.4.0
+docker==4.0.2
 docutils==0.14
 dogpile.cache==0.7.1
-filelock==3.0.10
+execnet==1.6.0
+filelock==3.0.12
 geojson==2.4.1
-girder==3.0.0a7.dev2
-girder-client==2.4.0
-girder-gravatar==3.0.0a7.dev2
-girder-jobs==3.0.0a7.dev2
-girder-large-image==1.0.0.dev31
+girder==3.0.0rc2.dev121
+girder-client==3.0.0rc2.dev121
+girder-gravatar==3.0.0rc2.dev121
+girder-jobs==3.0.0rc2.dev121
+girder-large-image==1.0.0.dev52
 girder-worker==0.6.0
-girder-worker-utils==0.8.4
+girder-worker-utils==0.8.6.dev2
 idna==2.8
 imageio==2.5.0
+importlib-metadata==0.18
 jaraco.functools==2.0
 jmespath==0.9.4
-jsonpickle==1.1
+jsonpickle==1.2
 jsonschema==3.0.1
-kiwisolver==1.0.1
-kombu==4.5.0
-large-image==1.0.0.dev31
-large-image-source-tiff==1.0.0.dev31
+kiwisolver==1.1.0
+kombu==4.6.3
+large-image==1.0.0.dev52
+large-image-source-tiff==1.0.0.dev52
 libtiff==0.4.2
-Mako==1.0.8
+Mako==1.0.12
 MarkupSafe==1.1.1
-matplotlib==3.0.3
+matplotlib==3.1.0
+mock==3.0.5
+mongomock==3.17.0
 more-itertools==7.0.0
 natsort==6.0.0
-networkx==2.2
-numpy==1.16.2
-opencv-python==4.0.1.24
+networkx==2.3
+numpy==1.16.4
+opencv-python==4.1.0.25
+packaging==19.0
 passlib==1.7.1
-pbr==5.1.3
+pbr==5.3.1
 Pillow==6.0.0
-pluggy==0.9.0
-portend==2.4
-psutil==5.6.1
+pluggy==0.12.0
+portend==2.5
+psutil==5.6.3
 py==1.8.0
 pycparser==2.19
-pymongo==3.7.2
+pymongo==3.8.0
 pyOpenSSL==19.0.0
 pyparsing==2.4.0
-pyrsistent==0.14.11
+pyrsistent==0.15.2
+pytest==4.6.3
+pytest-cov==2.7.1
+pytest-forked==1.0.2
+pytest-girder==3.0.0rc2.dev121
+pytest-xdist==1.29.0
 python-dateutil==2.8.0
-python-dotenv==0.10.1
+python-dotenv==0.10.3
 pytz==2019.1
 PyWavelets==1.0.3
-PyYAML==5.1
+PyYAML==5.1.1
 redis==3.2.1
-requests==2.21.0
+requests==2.22.0
 requests-toolbelt==0.9.1
-s3transfer==0.2.0
+s3transfer==0.2.1
 scikit-image==0.15.0
-scipy==1.2.1
-sentry-sdk==0.7.10
-setuptools-scm==3.2.0
+scipy==1.3.0
+sentinels==1.0.0
+sentry-sdk==0.9.1
+setuptools-scm==3.3.3
 six==1.12.0
 stevedore==1.30.1
-tempora==1.14
-toml==0.10.0
-tox==3.8.6
-ujson==1.35
-urllib3==1.24.1
+tempora==1.14.1
+urllib3==1.25.3
 vine==1.3.0
-virtualenv==16.4.3
+wcwidth==0.1.7
 websocket-client==0.56.0
+zipp==0.5.1
+tox
