+++ isic_archive/utility/mail_utils.py
@@ -19,9 +19,9 @@
 from email.mime.text import MIMEText
 
 from girder import logger
-from girder.constants import SettingKey
 from girder.models.group import Group
 from girder.models.setting import Setting
+from girder.settings import SettingKey
 from girder.utility import mail_utils
